Add password protection to be able to do stuff in vatpls.
- Started admin panel.
- Change website title to VATPLS.
